% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgp0a.R
\docType{package}
\name{bcgp0a}
\alias{bcgp0a}
\alias{bcgp0a-package}
\title{bcgp0a: a package for implementing Bayesian Composite Gaussian Processes}
\description{
The bcgp0a package performs MCMC and blah, blah, blah. Implemented purely in R.
It probably won't be fast at all, but I want to get something completed.
Also not well documented. It's basically a translation of the ugly MATLAB original
code.
}
\section{These are the functions}{

bcgp, MCMC, ...
}

