% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpfit-class.R
\docType{class}
\name{bcgpfit-class}
\alias{bcgpfit-class}
\title{Class \code{bcgpfit}: fitted bcgp model}
\description{
Class \code{bcgpfit}: fitted bcgp model
}
\section{Slots}{

\describe{
\item{\code{sim}}{A list containing simulation results including the
posterior draws as well as various pieces of metadata used by many of the
methods for \code{bcgpfit} objects.}

\item{\code{priors}}{A list containing the values of the prior parameters.}

\item{\code{inits}}{The initial values (either user-specified or generated
randomly) for all chains. This is a list with one component per chain. Each
component is a named list containing the initial values for each parameter
for the corresponding chain.}

\item{\code{algorithm}}{Either "Stan" or "M-H and Gibbs". Tells the method of sampling from the
posterior}
}}

\section{TODO}{
 Decide whether to put in \code{prototype} and \code{validity}
arguments.
}

