% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpMCMC.R
\name{bcgpMCMC}
\alias{bcgpMCMC}
\title{Draw samples from a bcgp model}
\usage{
bcgpMCMC(x, y, priors, inits, numUpdates, numAdapt, burnin, nmcmc,
  chains = 4, cores = 1)
}
\arguments{
\item{x}{An \code{n x d} matrix containing the independent variables
in the training set.}

\item{y}{A vector containing the observed response values in the training
set.}

\item{numUpdates}{The number of updates in the proposal stepsize adaptation phase.}

\item{numAdapt}{The number of samples within each update in the proposal stepsize
adaptation phase.}

\item{burnin}{The number of burnin samples to discard after the stepsize
adaptation phase is finished}

\item{nmcmc}{The number of samples to be kept for each Markov chain.}

\item{chains}{A positive integer specifying the number of Markov chains.
The default is 4.}

\item{cores}{The number of cores to use when executing the Markov chains in
parallel. The default is to use the value of the \code{mc.cores} option if it
has been set and otherwise to default to 1 core.}

\item{prior}{A list containing the values for the prior parameters.}
}
\value{
An object of S4 class \code{bcgp} representing the fitted results.
}
\description{
\code{bcgpMCMC} draws samples from the Bayesian Composite Gaussian Process model
}
\details{
This draws samples from the posterior distribution for the Bayesian
Composite Gaussian Process (BCGP) model.
}
\examples{

x <- matrix(runif(20, 0, 1), nrow = 10, ncol = 2)
y <- x[, 1] + sin(x[, 2])
priors <- createPriors(x, noise = FALSE)
inits <- createInits(x, priors, chains = 4)
numUpdates <- 3
numAdapt <- 500
burnin <- 500
nmcmc <- 5000
chains <- 4
cores <- 1
bcgpMCMC(x, y, priors, inits, numUpdates, numAdapt,
         burnin, nmcmc, chains, cores)
}
\seealso{
Other Major functions: \code{\link{bcgp}}
}
\concept{Major functions}
