% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineCorMats.R
\name{combineCorMats}
\alias{combineCorMats}
\title{Combine correlation matrices.}
\usage{
combineCorMats(w, G, L)
}
\arguments{
\item{w}{A scalar}

\item{G}{An \emph{n x n} correlation matrix, often the result of \code{\link{getCorMat}}}

\item{L}{An \emph{n x n} correlation matrix, often the result of \code{\link{getCorMat}}}
}
\value{
An \emph{n x n} correlation matrix
}
\description{
\code{combineCorMats} returns a correlation matrix that is a weighted sum of two
other correlation matrices.
}
\details{
This creates a correlation matrix, \emph{R}, where \deqn{R = wG + (1-w)L}In
the \code{bcgp} setting, \emph{G} is the global correlation matrix, \emph{L}
is the local correlation matrix, and \emph{w} is the weight.
}
\examples{
n <- 10
d <- 2
x <- matrix(runif(n * d), nrow = n, ncol = d)
rhoG <- runif(d, 0, 1)
rhoL <- runif(d, 0, rhoG)
G <- getCorMat(x, rhoG)
L <- getCorMat(x, rhoL)
w <- runif(1, 0.5, 1)
combineCorMats(w, G, L)
}
\seealso{
Other correlation and covariance functions: \code{\link{getCorMat}}
}
\concept{correlation and covariance functions}
