% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCovMat.R
\name{getCovMat}
\alias{getCovMat}
\title{Create a covariance matrix.}
\usage{
getCovMat(V, R, sig2eps)
}
\arguments{
\item{V}{A positive vector of length \emph{n}.}

\item{R}{An \emph{n x n} correlation matrix.}

\item{sig2eps}{A positive scalar representing the noise.}
}
\value{
An \emph{n x n} covariance matrix
}
\description{
\code{getCovMat} returns a covariance matrix (with nugget).
}
\details{
This creates a covariance matrix, \emph{C}, where
\deqn{C =  V^{0.5}RV^{0.5} + \sigma^2_\epsilon I}
where V is a matrix with variances on the diagonal. In the \code{bcgp}
setting, \emph{R} is a correlation matrix resulting from
\code{\link{combineCorMats}}, \emph{V} is a vector of process variances,
and \emph{sig2eps} is the variance of the noise.
}
\examples{
n <- 10
d <- 2
x <- matrix(runif(n * d), nrow = n, ncol = d)
rho <- runif(d, 0, 1)
R <- getCorMat(x, rho)
sig2eps <- 0.01
V <- rlnorm(n, -0.1, 0.1)
getCovMat(V, R, sig2eps)
}
\seealso{
Other correlation and covariance functions: \code{\link{combineCorMats}},
  \code{\link{getCorMat}}
}
\concept{correlation and covariance functions}
