% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logDet.R
\name{logDet}
\alias{logDet}
\title{Calculate the log of the determinant of a matrix.}
\usage{
logDet(C)
}
\arguments{
\item{C}{An \emph{n x n} covariance matrix.}
}
\value{
A scalar
}
\description{
\code{logDet} returns the log of the determinant of a matrix..
}
\details{
This calculates the log of the determinant of a matrix. In the \code{bcgp}
setting, \emph{C} is a covariance matrix resulting from
\code{\link{getCovMat}}.
}
\examples{
n <- 10
d <- 2
x <- matrix(runif(n * d), nrow = n, ncol = d)
rho <- runif(d, 0, 1)
R <- getCorMat(x, rho)
sig2 <- 0.01
V <- rlnorm(n, -0.1, 0.1)
C <- getCovMat(V, R, sig2)
logDet(C)
}
