% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logPost.R
\name{logPost}
\alias{logPost}
\title{Calculate the log posterior density.}
\usage{
logPost(x, y, params, priors, C, K)
}
\arguments{
\item{x}{An \code{n x d} matrix containing the independent variables
in the training set.}

\item{y}{A vector containing the observed response values in the training
set.}

\item{params}{A named vector containing the current value of each of
the parameters.}

\item{priors}{A named vector containing the prior information. In the
\code{bcgp} setting this will generally result from \code{unlist(priors)}.}

\item{C}{An \emph{n x n} covariance matrix for the training data.}

\item{K}{An \emph{n x n} covariance matrix for the variance process.
at the training data locations.}
}
\value{
A scalar
}
\description{
\code{logPost} returns the log of the posterior density up to a constant.
}
\details{
This calculates the log of the posterior density in the \code{bcgp}
setting up to a constant.
}
\examples{
x <- matrix(runif(20, 0, 1), nrow = 10, ncol = 2)
y <- x[, 1] + sin(x[, 2])
priors <- createPriors(x, noise = FALSE)
inits <- createInits(x, priors, chains = 1)
G <- getCorMat(x, inits[[1]]$rhoG)
L <- getCorMat(x, inits[[1]]$rhoL)
R <- combineCorMats(inits[[1]]$w, G, L)
C <- getCovMat(inits[[1]]$V, R, inits[[1]]$sig2eps)
K <- inits[[1]]$sig2V * getCorMat(x, inits[[1]]$rhoV) + diag(1e-10, length(y))
params <- unlist(inits[[1]])
priors <- unlist(priors)
logPost((x, y, params, priors, C, K))
}
